requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-ckeditor==6.7.1
whitenoise==6.6.0
gunicorn==21.2.0
